feat(verifier): add requiredClaims option
Before this commit, the verifier would not check for the existence of the claims
defined as allowed using the `allowed*` option, such as `allowedAud`.
This commit introduces a new option, `requiredClaims`, which can be used for
defining which claims should exist in the token.
Note that no claim is required by default, in order to preserve the behavior of
skipping the validation of undefined claims.