Include invalid claim value(s) in error messages
When some of the claims are rejected as not allowed:

Before my change, the error message described the claim name, but not
the value that's not allowed. As a result, it is difficult to
troubleshoot authentication errors.

    The aud claim value is not allowed.

In this commit, I have modified the code building the error message
to include the claim value extracted from the JWT token.

    The aud claim value is not allowed: "https://google.com"
    The aud claim value is not allowed: ["my-audience", "audience2"]

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>